Move TemplateModel class into separate file
