+++ mira/dkg/model.py
@@ -24,8 +24,9 @@ from mira.examples.sir import sir_bilayer, sir
 from mira.metamodel import NaturalConversion, Template, ControlledConversion
 from mira.metamodel.ops import stratify
 from mira.modeling import Model
-from mira.metamodel.templates import TemplateModel, TemplateModelDelta, \
+from mira.metamodel.templates import TemplateModelDelta, \
     Concept, Parameter, ModelComparisonGraphdata
+from mira.metamodel.template_model import TemplateModel
 from mira.metamodel.ops import simplify_rate_laws, aggregate_parameters
 from mira.modeling.bilayer import BilayerModel
 from mira.modeling.petri import PetriNetModel, PetriNetResponse
