+++ mira/examples/jin2022.py
@@ -7,7 +7,8 @@
 """
 
 from mira.metamodel import NaturalConversion
-from mira.metamodel.templates import ControlledConversion, TemplateModel
+from mira.metamodel.templates import ControlledConversion
+from mira.metamodel.template_model import TemplateModel
 from.concepts import susceptible, infected, recovered, exposed, dead
 
 __all__ = [
