+++ mira/metamodel/__init__.py
@@ -1,3 +1,4 @@
 # We expose everything that these submodules expose
 from .io import model_from_json_file, model_to_json_file
 from .templates import *
+from .template_model import TemplateModel
