+++ tests/test_model_api.py
@@ -15,8 +15,9 @@ from mira.dkg.web_client import is_ontological_child_web, get_relations_web
 from mira.metamodel import Concept, ControlledConversion, NaturalConversion, \
     RefinementClosure
 from mira.metamodel.ops import stratify
-from mira.metamodel.templates import TemplateModel, TemplateModelDelta, \
+from mira.metamodel.templates import TemplateModelDelta, \
     SympyExprStr, TemplateModelComparison, ModelComparisonGraphdata
+from mira.metamodel.template_model import TemplateModel
 from mira.modeling import Model
 from mira.modeling.bilayer import BilayerModel
 from mira.modeling.petri import PetriNetModel, PetriNetResponse
