+++ tests/test_modeling/test_ode.py
@@ -4,7 +4,8 @@ import unittest
 
 import numpy
 
-from mira.metamodel import TemplateModel, NaturalConversion, ControlledConversion, Concept
+from mira.metamodel import NaturalConversion, ControlledConversion, Concept
+from mira.metamodel.template_model import TemplateModel
 from mira.modeling import Model, Initial
 from mira.modeling.ode import OdeModel, simulate_ode_model
