+++ tests/test_template.py
@@ -3,7 +3,8 @@ import sympy
 from mira.metamodel import ControlledConversion, Concept, NaturalConversion, \
     NaturalDegradation, Template, GroupedControlledConversion, \
     GroupedControlledProduction
-from mira.metamodel.templates import Config, TemplateModel
+from mira.metamodel.templates import Config
+from mira.metamodel.template_model import TemplateModel
 from mira.dkg.web_client import is_ontological_child_web
 
 # Provide to tests that are not meant to test ontological refinements;
